dhraviq-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
firebase-admin
sendgrid
openai
requests


